app/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
plex.py==0.9.0
PlexAPI==4.0.0
pycparser==2.20
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
typing-extensions==3.7.4.2
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.1
yarl==1.4.2
